pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arcade>=2.3.15",
    "bokeh>=3.7.3",
    "bs4>=0.0.2",
    "colorama>=0.4.6",
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "google-genai>=1.29.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.34.4",
    "jax>=0.7.1",
    "kivy>=2.3.1",
    "langchain[google-genai]>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "matplotlib>=3.10.5",
    "mcp[cli]>=1.12.3",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "ollama>=0.5.2",
    "openai>=1.99.9",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pyfiglet>=1.0.3",
    "pygame>=2.6.1",
    "pyinstaller>=6.14.2",
    "pynsist>=2.8",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "scrapy>=2.13.3",
    "streamlit>=1.49.1",
    "taichi>=1.7.4",
    "tensorflow>=2.20.0",
    "termcolor>=3.1.0",
    "terminaltexteffects>=0.12.0",
    "textual>=4.0.0",
    "transformers>=4.55.2",
    "typer>=0.16.0",
    "unsloth>=2025.8.9",
    "ursina>=8.1.1",
    "uvicorn>=0.35.0",
    "wxpython>=4.2.3",
    "seaborn>=0.13.2",
    "langchain-openai>=0.3.32",
    "plotly>=6.3.0",
]
